Move DebuggerSidebar into its own file
